qemu: template packer config and install ILB archive
Install the Isovalent Load Balancer archive into a well-known directory
in the image.

Invoke the 'prepare_base.sh' script to install our base set of tools
needed for the Load Balancer installation.

Update the kernel using the upstream 'kernel-team' PPA. This includes
our fixes to XDP for the vmxnet3 driver. This is useful for local
debugging on KVM as well using this driver. It also keeps the kernel in
sync with the OVA image (as long as a push to this PPA doesn't happen
exactly at build time of our images).

Create envsubst placeholders in the packer config.

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>